// src/app/openapi/model/film.ts
/**
 * Filmkritiken-Backend
 * Filmkritiken-Backend
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Image } from './image';


export interface Film { 
    titel: string;
    altersfreigabe?: number;
    erscheinungsjahr?: number;
    regie?: string;
    laenge?: number;
    originaltitel?: string;
    originalsprache?: string;
    produktionsland?: string;
    image?: Image;
}